Update development instructions regarding figaro
If you don't have the correct Ruby version installed, per the
.ruby-version file, then `bundle exec figaro install` will
blow up with a _version not installed_ error.

If you take off the bundle exec, then it tries to run figaro,
which you may not have at this point.

We could update the instructions to say

    gem install figaro && figaro install

however that will generate a config/application.yml with
commented out environment variables in it, none of which we need.

There is a config/application.example.yml file which lists the
variables that we do need.

To bypass all the potential errors of trying to work with a gem
before running bootstrap and setup, this instructs people to
copy the example yaml and fill it out according to the tokens
generated for the oauth app.