user: correctly set OAuth token.
As noted in #66 users OAuth tokens were not being set correctly. This is
because they were using `auth.info.token` rather than
`auth.credentials.token` to get them back from GitHub/Devise. Fix this
up and additionally set the token for those users that are already
stored in the database but are missing this attribute.