Improve Rails cache utilisation
Instead of using a split approach between Octokit caching and Rails
caching cache the processed metadata from the Octokit requests.
This has the benefits of avoiding reprocessing on every request and
not storing in the cache data that we don't need for user display.

This involves changes to the cache keys and invalidating all existing
cache entries but this will be better for long-term resource usage
after the initial public ship.